Move util to specific package location
